1. Define CpuOut and its Bundle instance 2. Connect to dataRam
---

This is a 110 euro commit!

Note: readNew creates "read after write" behavior, however, it also
dramatically reduces the RAM needed for synthesis (or rather, the
version without it would produce Verilog that will consume all the RAM
when doing synthesis).

Another 2x8G RAM was purchased today because of the OOM problem with
synthesis, totaling 24G of RAM, and without 'readNew' the synthesis
would still take up %MEM to almost 100%, as well as almost all of the
8G swap. 'readNew' reduces %MEM to around 55%. I think this is a
problem with the Verilog generated. **And other seemingly trivial change
might also result in this dramatic difference.**